+++ lua/kubectl/init.lua
@@ -1,4 +1,3 @@
-local commands = require("kubectl.actions.commands")
 local informer = require("kubectl.actions.informer")
 local state = require("kubectl.state")
 
@@ -61,25 +60,10 @@ function M.setup(options)
   vim.api.nvim_create_user_command("Kubectl", function(opts)
     local view = require("kubectl.views")
     local action = opts.fargs[1]
-    local ok, x_view
     if action == "get" then
       if #opts.fargs == 2 then
         local resource_type = opts.fargs[2]
-        local viewsTable = require("kubectl.utils.viewsTable")
-        for key, item in pairs(viewsTable) do
-          if vim.tbl_contains(item, resource_type) then
-            ok, x_view = pcall(require, "kubectl.views." .. key)
-            if ok then
-              break
-            end
-          end
-        end
-        if ok then
-          pcall(x_view.View)
-        else
-          view = require("kubectl.views.fallback")
-          view.View(nil, opts.fargs[2])
-        end
+        view.view_or_fallback(resource_type)
       else
         view.UserCmd(opts.fargs)
       end
