+++ lua/kubectl/views/fallback/init.lua
@@ -53,7 +53,7 @@ function M.View(cancellationToken, resource)
     definition.cmd = "curl"
   end
 
-  M.builder = ResourceBuilder:new(definition.resource):view(definition, cancellationToken)
+  M.builder = ResourceBuilder:new(definition.resource):view(definition, cancellationToken, { cmd = definition.cmd })
 end
 
 function M.Draw(cancellationToken)
